Move all tests to tests/ folder.
The folder is provided by Cargo and makes it easier to find tests.
